feat(invoice): refactor item quantity to integer and show validation error messages inline

+++ InvoiceSystem.Web/Domain/Entities/InvoiceItem.cs
@@ -4,7 +4,7 @@ public class InvoiceItem
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal Quantity { get; set; }
+    public int Quantity { get; set; }
     public decimal UnitPrice { get; set; }
     public decimal TotalPrice { get; set; }
